Move st_json test to new file
