+++ lib/tests/streamlit/streamlit_test.py
@@ -179,23 +179,6 @@ class StreamlitTest(unittest.TestCase):
 class StreamlitAPITest(DeltaGeneratorTestCase):
     """Test Public Streamlit Public APIs."""
 
-    def test_st_json(self):
-        """Test st.json."""
-        st.json('{"some": "json"}')
-
-        el = self.get_delta_from_queue().new_element
-        self.assertEqual(el.json.body, '{"some": "json"}')
-
-        # Test that an object containing non-json-friendly keys can still
-        # be displayed.  Resultant json body will be missing those keys.
-
-        n = np.array([1, 2, 3, 4, 5])
-        data = {n[0]: "this key will not render as JSON", "array": n}
-        st.json(data)
-
-        el = self.get_delta_from_queue().new_element
-        self.assertEqual(el.json.body, '{"array": "array([1, 2, 3, 4, 5])"}')
-
     def test_st_legacy_line_chart(self):
         """Test st._legacy_line_chart."""
         df = pd.DataFrame([[10, 20, 30]], columns=["a", "b", "c"])
